Downloads/Competitor-Analysis-Engine-main/requirements.txt
streamlit>=1.29.0
openai>=1.12.0
duckduckgo_search>=5.1.0
tavily-python>=0.3.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
openpyxl>=3.1.2
python-pptx>=0.6.21
fpdf>=1.7.2
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
